% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bitre_hard_coded}
\alias{get_bitre_hard_coded}
\title{Get BITRE data with hard coded link}
\usage{
get_bitre_hard_coded(group = "fatal_crashes")
}
\arguments{
\item{group}{Character. The group name.}
}
\value{
path to the temporary XLSX file.
}
\description{
Get BITRE data with hard coded link
}
