% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-fatal.R
\name{get_oz_fatal}
\alias{get_oz_fatal}
\title{Retrieve the data of fatalities from Australian Road Deaths Database}
\usage{
get_oz_fatal()
}
\description{
Retrieve the data of fatalities from Australian Road Deaths Database
}
\examples{
\dontrun{
get_oz_fatal()
}
}
