% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-clean.R
\name{get_oz_fatal_crash}
\alias{get_oz_fatal_crash}
\title{Get the australian fatal crash data}
\source{
\url{https://bitre.gov.au/statistics/safety/fatal_road_crash_database.aspx}
}
\usage{
get_oz_fatal_crash()
}
\description{
Get the australian fatal crash data
}
\examples{
\dontrun{
get_oz_fatal_crash()
}
}
