% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oz-road-fatal-crash.R
\name{oz_road_fatal_crash_bitre}
\alias{oz_road_fatal_crash_bitre}
\title{Retrieve road fatal crash data from BITRE}
\usage{
oz_road_fatal_crash_bitre()
}
\description{
Retrieve road fatal crash data from BITRE
}
