% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data-clean.R
\name{get_oz_fatal_crash}
\alias{get_oz_fatal_crash}
\title{Get the australian fatal crash data}
\usage{
get_oz_fatal_crash()
}
\value{
a dataset (tibble) of fatal crash data
}
\description{
Get the australian fatal crash data
}
\examples{
\dontrun{
get_oz_fatal_crash()
}
}
